requirements_dev.txt
pip==22.3.1
wheel==0.38.4
Pillow==9.3.0